refactor: 쿠폰과 쿠폰 그룹 테스트의 컨트롤러, 서비스 분리 (#171)
