+++ src/test/java/com/firstone/greenjangteo/coupon/service/CouponServiceTest.java
@@ -21,26 +21,20 @@ import org.junit.jupiter.api.Test;
 import org.springframework.batch.core.JobExecutionException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityNotFoundException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static com.firstone.greenjangteo.coupon.excpeption.message.NotFoundExceptionMessage.COUPON_GROUP_ID_NOT_FOUND_EXCEPTION;
-import static com.firstone.greenjangteo.coupon.excpeption.message.NotFoundExceptionMessage.COUPON_NAME_NOT_FOUND_EXCEPTION;
 import static com.firstone.greenjangteo.coupon.testutil.CouponTestConstant.*;
 import static com.firstone.greenjangteo.user.model.Role.ROLE_BUYER;
 import static com.firstone.greenjangteo.user.testutil.UserTestConstant.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
 
 @ActiveProfiles("test")
 @SpringBootTest
@@ -52,10 +46,10 @@ class CouponServiceTest {
     private AuthenticationService authenticationService;
 
     @Autowired
-    private CouponGroupRepository couponGroupRepository;
+    private CouponRepository couponRepository;
 
     @Autowired
-    private CouponRepository couponRepository;
+    private CouponGroupRepository couponGroupRepository;
 
     @Autowired
     private UserRepository userRepository;
@@ -233,84 +227,6 @@ class CouponServiceTest {
 
         // then
         assertThat(foundCoupons).hasSize(coupons2.size());
-    }
-
-    @DisplayName("쿠폰 그룹 ID를 통해 쿠폰 목록을 페이징 처리한 쿠폰 그룹을 찾을 수 있다.")
-    @Test
-    @Transactional
-    void getCouponGroupByCouponGroupId() {
-        // given
-        CouponGroup createdCouponGroup
-                = CouponTestObjectFactory.createCouponGroup(
-                COUPON_NAME1, AMOUNT, DESCRIPTION, ISSUE_QUANTITY1, tomorrow, EXPIRATION_PERIOD1
-        );
-        couponGroupRepository.save(createdCouponGroup);
-
-        List<Coupon> createdCoupons = CouponTestObjectFactory.createCoupons(createdCouponGroup);
-        couponRepository.saveAll(createdCoupons);
-
-        // when
-        List<Coupon> foundCoupons = couponService.getCouponGroup(createdCouponGroup.getId(), Pageable.ofSize(2))
-                .getContent();
-
-        // then
-        CouponGroup foundCouponGroup = foundCoupons.get(0).getCouponGroup();
-
-        assertThat(foundCouponGroup.getCouponName()).isEqualTo(createdCouponGroup.getCouponName());
-        assertThat(foundCouponGroup.getAmount()).isEqualTo(createdCouponGroup.getAmount());
-        assertThat(foundCouponGroup.getDescription()).isEqualTo(createdCouponGroup.getDescription());
-        assertThat(foundCouponGroup.getIssueQuantity()).isEqualTo(createdCouponGroup.getIssueQuantity());
-        assertThat(foundCouponGroup.getScheduledIssueDate()).isEqualTo(createdCouponGroup.getScheduledIssueDate());
-        assertThat(foundCouponGroup.getExpirationPeriod()).isEqualTo(createdCouponGroup.getExpirationPeriod());
-
-        assertThat(foundCoupons).hasSize(2);
-    }
-
-    @DisplayName("전송된 쿠폰 그룹 ID를 가진 쿠폰 그룹이 존재하지 않으면 EntityNotFoundException이 발생한다.")
-    @Test
-    @Transactional
-    void getCouponGroupByNonExistentCouponGroupId() {
-        // given
-        Long couponGroupId = 1L;
-        Pageable pageable = mock(Pageable.class);
-
-        // when, then
-        assertThatThrownBy(() -> couponService.getCouponGroup(couponGroupId, pageable))
-                .isInstanceOf(EntityNotFoundException.class)
-                .hasMessage(COUPON_GROUP_ID_NOT_FOUND_EXCEPTION + couponGroupId);
-    }
-
-    @DisplayName("쿠폰 이름을 통해 쿠폰 그룹을 찾을 수 있다.")
-    @Test
-    @Transactional
-    void getCouponGroupByCouponName() {
-        // given
-        CouponGroup createdCouponGroup
-                = CouponTestObjectFactory.createCouponGroup(
-                COUPON_NAME1, AMOUNT, DESCRIPTION, ISSUE_QUANTITY1, tomorrow, EXPIRATION_PERIOD1
-        );
-        couponGroupRepository.save(createdCouponGroup);
-
-        // when
-        CouponGroup foundCouponGroup = couponService.getCouponGroup(COUPON_NAME1);
-
-        // then
-        assertThat(foundCouponGroup.getCouponName()).isEqualTo(createdCouponGroup.getCouponName());
-        assertThat(foundCouponGroup.getAmount()).isEqualTo(createdCouponGroup.getAmount());
-        assertThat(foundCouponGroup.getDescription()).isEqualTo(createdCouponGroup.getDescription());
-        assertThat(foundCouponGroup.getIssueQuantity()).isEqualTo(createdCouponGroup.getIssueQuantity());
-        assertThat(foundCouponGroup.getScheduledIssueDate()).isEqualTo(createdCouponGroup.getScheduledIssueDate());
-        assertThat(foundCouponGroup.getExpirationPeriod()).isEqualTo(createdCouponGroup.getExpirationPeriod());
-    }
-
-    @DisplayName("전송된 쿠폰 이름을 가진 쿠폰 그룹이 존재하지 않으면 EntityNotFoundException이 발생한다.")
-    @Test
-    @Transactional
-    void getCouponGroupByNonExistentCouponName() {
-        // given, when, then
-        assertThatThrownBy(() -> couponService.getCouponGroup(COUPON_NAME1))
-                .isInstanceOf(EntityNotFoundException.class)
-                .hasMessage(COUPON_NAME_NOT_FOUND_EXCEPTION + COUPON_NAME1);
     }
 
     @DisplayName("미리 발급된 쿠폰을 전송된 수량만큼 회원에게 지급할 수 있다.")
